CSV add csv file -add city class in proper place
