+++ src/cityreader.py
@@ -2,8 +2,11 @@
 # fields for name, latitude, and longitude.
 
 # TODO
- def drive(self):
-        print("vrooom")
+ class City:
+    def __init__(self, name, lat, lon):
+        self.name = name
+        self.lat = lat
+        self.lon = lon
 
 # We have a collection of US cities with population over 750,000 stored in the
 # file "cities.csv". (CSV stands for "comma-separated values".)
@@ -21,12 +24,14 @@
 cities = []
 
 # TODO
-class Motorcycle(GroundVehicle):
-    def __init__(self, num_wheels=2):
-        super().__init__(num_wheels)
+import csv
 
-    def drive(self):
-        print("BRAAAP!!")
+with open("cities.csv", newline="") as csvfile:
+    reader = csv.reader(csvfile)
+    for c in reader:
+        cities.append(c)
+
+cities.pop(0)
 # Print the list of cities (name, lat, lon), 1 record per line.
 
 # TODO
