+++ src/oop2.py
@@ -4,10 +4,12 @@
 # object is constructed.
 
 class GroundVehicle():
-    def __init__(self, num_wheels):
+    def __init__(self, num_wheels=4):
         self.num_wheels = num_wheels
 
     # TODO
+    def drive(self):
+        print("vroooom")
 
 
 # Subclass Motorcycle from GroundVehicle.
@@ -18,6 +20,12 @@ class GroundVehicle():
 # Override the drive() method in Motorcycle so that it prints "BRAAAP!!"
 
 # TODO
+class Motorcycle(GroundVehicle):
+    def __init__(self, num_wheels=2):
+        super().__init__(num_wheels)
+
+    def drive(self):
+        print("BRAAAP!!")
 
 vehicles = [
     GroundVehicle(),
@@ -30,3 +38,5 @@ vehicles = [
 # Go through the vehicles list and call drive() on each.
 
 # TODO
+for vehicle in vehicles:
+    vehicle.drive()
